+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 1.0a2 (unreleased)
 ------------------
 
+- Render errors as JSON.
+  [jone]
+
 - Add support for named services which allows registering services like
   ``GET /Plone/search`` or ``GET /Plone/doc1/versions/1`` using a 'name' attribute.
   [jone, lukasgraf, buchi]